+++ aern2-mfun/src/AERN2/AD/MPBallOperations.hs
@@ -6,6 +6,7 @@ import MixedTypesNumPrelude
 import AERN2.AD.Type
 import AERN2.AD.GenericOperations ()
 import AERN2.MP.Ball
+import Numeric.CollectErrors.Type (noValueNumErrorCertain, NumError (NumError))
 
 {-- addition --}
 
@@ -210,3 +211,13 @@ instance
                 dx_ <- dx
                 pure $ (hullMPBall dx_ (-dx_))
     abs (OrderTwo _ _ _ _) = error "Abs for differential order two undefined"
+
+instance 
+    CanDivIMod (Differential (CN MPBall)) (Differential (CN MPBall))
+    where
+    type DivIType (Differential (CN MPBall)) (Differential (CN MPBall)) = (Differential (CN MPBall))
+    divIMod a b = (error "Integer division for Differential (CN MPBall) undefined", OrderTwo (mod ax bx) err err err)
+        where   
+            ax = diff_x a
+            bx = diff_x b
+            err = noValueNumErrorCertain $ NumError "No derivatives after modulus"