+++ aern2-mp/changelog.md
@@ -2,6 +2,7 @@
 
 * current
   * more instances for CN Kleenean
+  * Num MPFloat (unspecified rounding mode)
 * v 0.2.8 2021-08-04
   * compatibility with ghc 9.0.1
   * add limits for CN (WithCurrentPrec p (CN MPBall))